Merged: [api] Fix compilation error with gcc
Revision: 272e26af0203b4e33b17514aecf8bb2caa2ace3f

BUG=v8:12566
NOTRY=true
NOPRESUBMIT=true
NOTREECHECKS=true
R=mfarazma@redhat.com

Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/3403735
Commit-Queue: Milad Farazmand <mfarazma@redhat.com>
Cr-Commit-Position: refs/branch-heads/9.9@{#2}
Cr-Branched-From: c4208cca4b053300b7ae3d297790c7dcd0be551a-refs/heads/9.9.115@{#1}
Cr-Branched-From: 19c6fd735108efee9d782c8f01ce0b7611a93621-refs/heads/main@{#78682}

+++ src/init/v8.cc
@@ -70,14 +70,15 @@ void AdvanceStartupState(V8StartupState expected_next_state) {
     // isolate->Dispose();
     // v8::V8::Dispose();
     // v8::V8::DisposePlatform();
-    FATAL("Wrong intialization order: got %d expected %d!", current_state,
-          next_state);
+    FATAL("Wrong intialization order: got %d expected %d!",
+          static_cast<int>(current_state), static_cast<int>(next_state));
   }
   if (!v8_startup_state_.compare_exchange_strong(current_state, next_state)) {
     FATAL(
         "Multiple threads are initializating V8 in the wrong order: expected "
         "%d got %d!",
-        current_state, v8_startup_state_.load());
+        static_cast<int>(current_state),
+        static_cast<int>(v8_startup_state_.load()));
   }
 }
